+++ src/main/java/com/nguyenquyhy/discordbridge/DiscordBridge.java
@@ -33,7 +33,7 @@ import java.util.*;
 /**
  * Created by Hy on 1/4/2016.
  */
-@Plugin(id = "discordbridge", name = "Discord Bridge", version = "2.2.0",
+@Plugin(id = "discordbridge", name = "Discord Bridge", version = "2.3.0",
         description = "A Sponge plugin to connect your Minecraft server with Discord", authors = { "Hy" })
 public class DiscordBridge {